+++ include/ddc/kernels/splines.hpp
@@ -11,9 +11,6 @@
 #include "splines/greville_interpolation_points.hpp"
 #include "splines/knots_as_interpolation_points.hpp"
 #include "splines/math_tools.hpp"
-#include "splines/matrix.hpp"
-#include "splines/matrix_maker.hpp"
-#include "splines/matrix_sparse.hpp"
 #include "splines/null_extrapolation_rule.hpp"
 #include "splines/periodic_extrapolation_rule.hpp"
 #include "splines/spline_boundary_conditions.hpp"
@@ -21,4 +18,7 @@
 #include "splines/spline_builder_2d.hpp"
 #include "splines/spline_evaluator.hpp"
 #include "splines/spline_evaluator_2d.hpp"
+#include "splines/splines_linear_solver.hpp"
+#include "splines/splines_linear_solver_maker.hpp"
+#include "splines/splines_linear_solver_sparse.hpp"
 #include "splines/view.hpp"
